Move grabber code to a separate file
